+++ src/plugins/migration/data-migrator.ts
@@ -541,6 +541,7 @@ export async function _migrateDocuments(
         const writeDeleted = flatClone(docData);
         writeDeleted._deleted = true;
         writeDeleted._attachments = {};
+        writeDeleted._rev = createRevision(writeDeleted, docData);
         return {
             previous: docData,
             document: writeDeleted