+++ test/unit/rx-collection.test.ts
@@ -1,7 +1,12 @@
 import assert from 'assert';
 import clone from 'clone';
 import config from './config';
-import AsyncTestUtil, { randomBoolean, randomNumber, randomString, wait } from 'async-test-util';
+import AsyncTestUtil, {
+    randomBoolean,
+    randomNumber,
+    randomString,
+    wait
+} from 'async-test-util';
 
 import * as schemas from '../helper/schemas';
 import * as schemaObjects from '../helper/schema-objects';
@@ -24,7 +29,6 @@ import {
     ensureNotFalsy,
     lastOfArray,
     now,
-    RxDocument,
     getFromMapOrThrow,
     RxCollectionCreator,
     parseRevision
@@ -1104,6 +1108,8 @@ config.parallel('rx-collection.test.js', () => {
                 it('should remove all documents', async () => {
                     const c = await humansCollection.create(10);
                     const query = c.find();
+
+
                     const removed = await query.remove();
                     assert.strictEqual(removed.length, 10);
                     removed.forEach(doc => {
@@ -1117,6 +1123,7 @@ config.parallel('rx-collection.test.js', () => {
                 it('should remove only found documents', async () => {
                     const c = await humansCollection.create(10);
                     const query = c.find().limit(5);
+
                     const removed = await query.remove();
                     assert.strictEqual(removed.length, 5);
                     removed.forEach(doc => {
@@ -1124,8 +1131,12 @@ config.parallel('rx-collection.test.js', () => {
                         assert.strictEqual(doc.deleted, true);
                     });
                     const docsAfter = await c.find().exec();
+                    console.dir(docsAfter.map(d => d.toJSON(true)));
                     assert.strictEqual(docsAfter.length, 5);
                     c.database.destroy();
+
+
+                    // process.exit();
                 });
                 it('remove on findOne', async () => {
                     const c = await humansCollection.create(10);
@@ -1473,6 +1484,7 @@ config.parallel('rx-collection.test.js', () => {
                     });
                     const collection = collections.human;
                     const obj = schemaObjects.simpleHuman();
+
                     await collection.insert(obj);
                     obj.firstName = 'foobar';
                     await collection.upsert(obj);
@@ -1498,6 +1510,7 @@ config.parallel('rx-collection.test.js', () => {
                     const objData = schemaObjects.simpleHuman();
 
                     const doc = await collection.insert(objData);
+
                     await doc.atomicPatch({
                         firstName: 'alice'
                     });
@@ -1638,10 +1651,9 @@ config.parallel('rx-collection.test.js', () => {
                         c.atomicUpsert(docData)
                     ]);
 
-
                     const viaStorage = await c.storageInstance.findDocumentsById([docId], true);
                     const viaStorageDoc = viaStorage[docId];
-                    assert.strictEqual(parseRevision(viaStorageDoc._rev).height, 3);
+                    assert.ok(parseRevision(viaStorageDoc._rev).height >= 3);
 
                     const docData2 = clone(docData);
                     docData2.firstName = 'foobar';
@@ -2215,7 +2227,7 @@ config.parallel('rx-collection.test.js', () => {
             const collection = await humansCollection.create(0);
 
             //  Record subscription
-            const emitted: Map<string, RxDocument<HumanDocumentType>>[] = [];
+            const emitted: Map<string, RxDocumentData<HumanDocumentType>>[] = [];
 
             function createObject(id: string): RxDocumentData<HumanDocumentType> {
                 const ret: RxDocumentData<HumanDocumentType> = Object.assign(
@@ -2236,7 +2248,14 @@ config.parallel('rx-collection.test.js', () => {
             const matchingIds = ['a', 'b', 'c', 'd'];
 
             const sub = collection.findByIds$(matchingIds).subscribe(data => {
-                emitted.push(data);
+
+                const m = new Map();
+                Array
+                    .from(data.entries())
+                    .forEach(([id, doc]) => {
+                        m.set(id, doc.toJSON(true));
+                    });
+                emitted.push(m);
             });
 
             //  test we have a map and no error
@@ -2277,28 +2296,31 @@ config.parallel('rx-collection.test.js', () => {
 
 
             /**
-             * Each emitted result must have a higher amount of documents,
+             * Each emitted result must have a different result set
              * because findByIds$ must only emit when data has actually changed.
              * We cannot just cound the updates.length here because some RxStorage implementations
-             * might return multiple RxChangeEventBulks for a single bulkWrite() operation.
+             * might return multiple RxChangeEventBulks for a single bulkWrite() operation
+             * or do additional writes. So we have to check for the revisions+docId strings.
              */
-            let lastCount: number;
+            const resultIds = new Set<string>();
+            console.dir(emitted);
             emitted.forEach(oneResult => {
-                if (typeof lastCount === 'undefined') {
-                    lastCount = oneResult.size;
+                let resultId = '';
+                Array.from(oneResult.entries()).forEach(([docId, docData]) => {
+                    resultId += docId + '|' + docData._rev + '-';
+                });
+                if (resultIds.has(resultId)) {
+                    throw new Error('duplicate result ' + resultId);
                 } else {
-                    if (oneResult.size <= lastCount) {
-                        throw new Error('emitted data not newer ' + oneResult.size + '- lastCount: ' + lastCount);
-                    }
-                    lastCount = oneResult.size;
+                    resultIds.add(resultId);
                 }
             });
 
             // should have the same result set as running findByIds() once.
             const singleQueryDocs = await collection.findByIds(matchingIds);
 
-            const singleResultPlain = matchingIds.map(id => getFromMapOrThrow(singleQueryDocs, id).toJSON())
-            const observedResultPlain = matchingIds.map(id => getFromMapOrThrow(lastOfArray(emitted), id).toJSON())
+            const singleResultPlain = matchingIds.map(id => getFromMapOrThrow(singleQueryDocs, id).toJSON(true))
+            const observedResultPlain = matchingIds.map(id => getFromMapOrThrow(lastOfArray(emitted), id))
             assert.deepStrictEqual(singleResultPlain, observedResultPlain);
 
             //  And contains the right data