+++ test/unit/rx-document.test.ts
@@ -439,7 +439,12 @@ config.parallel('rx-document.test.js', () => {
                 c.database.destroy();
             });
             it('should work when inserting on a slow storage', async () => {
-                if (!config.platform.isNode()) return;
+                if (
+                    !config.platform.isNode() ||
+                    config.storage.name !== 'pouchdb'
+                ) {
+                    return;
+                }
                 // use a 'slow' adapter because memory might be to fast
                 const leveldown = require('leveldown');
                 const db = await createRxDatabase({
@@ -454,20 +459,35 @@ config.parallel('rx-document.test.js', () => {
                 const c = cols.humans;
                 await c.insert(schemaObjects.simpleHuman());
                 const doc = await c.findOne().exec();
+
+                console.log('xxxx 1');
+
                 doc.atomicUpdate((innerDoc: any) => {
                     innerDoc.firstName = 'foobar';
                     return innerDoc;
                 });
+
+                console.log('xxxx 2');
+
                 await doc.atomicUpdate((innerDoc: any) => {
                     innerDoc.firstName = 'foobar2';
                     return innerDoc;
                 });
+
+                console.log('xxxx 3');
+
                 await AsyncTestUtil.wait(50);
+
+                console.log('xxxx 4');
+
                 await doc.atomicUpdate((innerDoc: any) => {
                     innerDoc.firstName = 'foobar3';
                     return innerDoc;
                 });
                 assert.strictEqual('foobar3', doc.firstName);
+
+                console.log('xxxx 5');
+
 
                 db.destroy();
             });