+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Only check if final field have been changed in dev-mode.
 - Fix `atomicUpsert()` broken when document was replicated before. [#3856](https://github.com/pubkey/rxdb/pull/3856). Thanks [@AntonOfTheWoods](https://github.com/AntonOfTheWoods)
+- Refactor revision handling
 
 <!-- ADD new changes here! -->